# contracts/api.yaml
openapi: 3.1.0
info:
  title: Kiezmarkt Listing Service
  version: 0.1.0
  summary: Listings and their lifecycle, search, categories, saved searches.
  description: |
    The listing service surface of Kiezmarkt, the synthetic neighbourhood
    classifieds marketplace used as the guiding project of the AI-enablement
    training.

    The domain model behind this contract is `docs/project/domain.md`. Where the
    two disagree, `domain.md` wins and this file is the defect. Two invariants
    are carried by the schemas rather than by prose:

    - Money is an integer number of cents. No property in this document is typed
      `number`, at any depth.
    - No response schema exposes a seller email. A seller is visible as
      `sellerId`, and where a human-readable name is required, through the
      `SellerPublic` projection.

    `domain.md` specifies no authentication or authorization model. Endpoints
    that are inherently caller-scoped (`/saved-searches`) say so in their
    description; the mechanism is out of scope for this contract.
servers:
  - url: https://api.kiezmarkt.example/v1
    description: Synthetic. Nothing is deployed behind this URL.
...
paths:
...
  /saved-searches:

    get:
      operationId: listSavedSearches
      tags: [saved-searches]
      summary: The caller's saved searches
      description: |
        Scoped to the owning account. A caller never sees another account's
        saved searches. `domain.md` defines no authentication model, so the
        mechanism that establishes the caller is out of scope here; the scoping
        rule is not.

        There is no pagination. `domain.md` sets no upper bound on saved
        searches per account; the seed set has 12 in total.
      responses:
        '200':
          description: All saved searches owned by the caller.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavedSearch'
              examples:
                two:
                  value:
                    - id: ss_4b1e07
                      sellerId: sel_0142
                      label: Plattenspieler bis 300
                      query: plattenspieler
                      filters:
                        categoryId: elektronik.audio
                        priceMinCents: 0
                        priceMaxCents: 30000
                        postcode: '10437'
                        radiusKm: 10
                      notify: true
                      lastMatchedAt: '2026-08-28T06:12:19Z'
                    - id: ss_88c210
                      sellerId: sel_0142
                      label: Kostenlose Moebel Neukoelln
                      query: ''
                      filters:
                        categoryId: haushalt.moebel
                        priceMinCents: 0
                        priceMaxCents: 0
                        postcode: '12047'
                        radiusKm: 5
                      notify: false
                      lastMatchedAt: null

    post:
      operationId: createSavedSearch
      tags: [saved-searches]
      summary: Save a search
      responses:
        '201':
          description: Created.
          headers:
            Location:
              description: URI of the created saved search.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedSearch'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedSearchCreate'
            examples:
              notifying:
                summary: A search that notifies on new matches
                value:
                  label: Plattenspieler bis 300
                  query: plattenspieler
                  filters:
                    categoryId: elektronik.audio
                    priceMaxCents: 30000
                    postcode: '10437'
                    radiusKm: 10
                  notify: true
  /saved-searches/{id}:

    parameters:
      - $ref: '#/components/parameters/SavedSearchId'

    patch:
      operationId: updateSavedSearch
      tags: [saved-searches]
      summary: Partially update a saved search
      description: |
        Absent properties are left unchanged. `filters` is replaced as a whole,
        not merged: sending `filters` with only `postcode` clears the other
        filter fields. `id`, `sellerId` and `lastMatchedAt` are server-owned and
        rejected in a patch body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedSearchPatch'
            examples:
              rename:
                summary: Rename
                value:
                  label: Plattenspieler, Prenzlauer Berg
              muteNotifications:
                summary: Stop notifying
                value:
                  notify: false
      responses:
        '200':
          description: The updated saved search.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedSearch'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      operationId: deleteSavedSearch
      tags: [saved-searches]
      summary: Delete a saved search
      description: |
        Hard delete. Unlike a listing, a saved search has no `deleted` status and
        leaves no tombstone in the API. Deleting a search stops its
        `savedsearch.matched` events from the next match evaluation on.
      responses:
        '204':
          description: Deleted. No body.
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
...
    SavedSearchId:
      name: id
      in: path
      required: true
      description: Opaque, server-assigned saved-search id.
      schema:
        $ref: '#/components/schemas/SavedSearchId'
  responses:
    BadRequest:
      description: The request is malformed or violates a field constraint.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            titleTooShort:
              value:
                type: https://kiezmarkt.example/problems/validation-failed
                title: Validation failed
                status: 400
                detail: title must be between 3 and 80 characters, got 2.
            floatPrice:
              summary: Money sent as a float
              value:
                type: https://kiezmarkt.example/problems/validation-failed
                title: Validation failed
                status: 400
                detail: priceCents must be an integer number of cents, got 25.99.
            radiusWithoutPostcode:
              value:
                type: https://kiezmarkt.example/problems/radius-without-postcode
                title: Radius filter needs a postcode
                status: 400
                detail: radiusKm was supplied without postcode.
            badCursor:
              value:
                type: https://kiezmarkt.example/problems/invalid-cursor
                title: Invalid cursor
                status: 400
                detail: The cursor is malformed or no longer valid.
    NotFound:
      description: No such resource, or the caller may not see it.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            notFound:
              value:
                type: https://kiezmarkt.example/problems/not-found
                title: Not found
                status: 404
                detail: No listing with id lst_000000.
...
    UnprocessableEntity:
      description: |
        Syntactically valid, but rejected by a domain rule: an unknown
        `categoryId`, a top-level `categoryId` on a listing (invariant 4), an
        unknown `sellerId`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            parentCategory:
              value:
                type: https://kiezmarkt.example/problems/category-not-a-leaf
                title: Category is not a leaf
                status: 422
                detail: elektronik has children and cannot carry listings directly.
  schemas:
    MoneyCents:
      title: MoneyCents
      description: |
        An amount of money as an integer number of cents. Never a float, at any
        layer (invariant 3 in `domain.md`). `currency` is carried separately.
      type: integer
      format: int64
      minimum: 0
      examples: [42000, 0, 2500]
    Postcode:
      title: Postcode
      description: Five-digit German postcode. Leading zeros are significant, so this is a string.
      type: string
      pattern: '^[0-9]{5}$'
      examples: ['10437', '12047', '13355']
...
    SellerId:
      type: string
      minLength: 1
      maxLength: 64
      examples: [sel_0142]
    SavedSearchId:
      type: string
      minLength: 1
      maxLength: 64
      examples: [ss_4b1e07]
    CategoryId:
      description: |
        Current form, dot-separated, at most two levels: `elektronik`,
        `elektronik.audio`. The pre-2024 numeric id is `legacyId` and is never
        accepted here.
      type: string
      pattern: '^[a-z0-9]+(\.[a-z0-9]+)?$'
      examples: [elektronik.audio, haushalt.moebel]
...
    SavedSearchFilters:
      title: SavedSearchFilters
      description: |
        The filter half of a saved search. Field for field the same filters as
        the `GET /listings` query parameters. Every field is optional; an empty
        object means the free-text `query` alone decides.
      type: object
      additionalProperties: false
      properties:
        categoryId:
          $ref: '#/components/schemas/CategoryId'
        priceMinCents:
          $ref: '#/components/schemas/MoneyCents'
        priceMaxCents:
          $ref: '#/components/schemas/MoneyCents'
        postcode:
          $ref: '#/components/schemas/Postcode'
        radiusKm:
          type: integer
          minimum: 1
          description: Requires `postcode`.
    SavedSearch:
      title: SavedSearch
      type: object
      additionalProperties: false
      required:
        - id
        - sellerId
        - label
        - query
        - filters
        - notify
        - lastMatchedAt
      properties:
        id:
          allOf: [{ $ref: '#/components/schemas/SavedSearchId' }]
          readOnly: true
        sellerId:
          allOf: [{ $ref: '#/components/schemas/SellerId' }]
          description: |
            The owner. Buyers and sellers are one account type, so a saved search
            is owned by a `sellerId` like a listing is.
          readOnly: true
        label:
          type: string
          minLength: 1
          maxLength: 80
          description: User-supplied and renameable.
        query:
          type: string
          maxLength: 200
          description: Free text. May be empty when the filters carry the search.
        filters:
          $ref: '#/components/schemas/SavedSearchFilters'
        notify:
          type: boolean
          description: |
            Whether a new published match emits `savedsearch.matched`. See
            `docs/project/contracts/events.md`.
        lastMatchedAt:
          type: [string, 'null']
          format: date-time
          description: UTC. Null until the first match.
          readOnly: true
    SavedSearchCreate:
      title: SavedSearchCreate
      type: object
      additionalProperties: false
      required:
        - label
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 80
        query:
          type: string
          maxLength: 200
          default: ''
        filters:
          $ref: '#/components/schemas/SavedSearchFilters'
        notify:
          type: boolean
          default: false
    SavedSearchPatch:
      title: SavedSearchPatch
      description: |
        Every property optional; absent means unchanged. `filters` is replaced
        wholesale, not merged.
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 80
        query:
          type: string
          maxLength: 200
        filters:
          $ref: '#/components/schemas/SavedSearchFilters'
        notify:
          type: boolean

    # The only permitted public shape of a seller. `email` is absent by
    # construction, not by filtering (invariant 6). No endpoint in this contract
    # returns it yet: `domain.md` defines no seller endpoint, and the schema is
    # kept here so that any future one has a shape to reuse rather than invent.
...
    Problem:
      title: Problem
      description: |
        RFC 9457 problem details. Every error response in this contract uses this
        shape with media type `application/problem+json`.
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: |
            Stable identifier of the problem kind. Clients branch on this, never
            on `detail`.
          default: about:blank
          examples:
            - https://kiezmarkt.example/problems/illegal-status-transition
        title:
          type: string
          description: Short, human-readable, constant per `type`.
        status:
          type: integer
          minimum: 400
          maximum: 599
          description: Repeats the HTTP status code.
        detail:
          type: string
          description: |
            Human-readable explanation of this occurrence. Free text. Never
            carries a seller email or any other PII.
        instance:
          type: string
          format: uri-reference
          description: URI of the specific occurrence.